fix: pass ruff lint + mypy strict type checks for CI
ruff:
- Add per-file-ignores for B008 (FastAPI Depends in defaults) and B017
  (pytest.raises(Exception) in tests)
- Fix UP042: str+Enum → enum.StrEnum (5 classes)
- Fix E501: break 14 long lines across 8 files
- Fix F841: remove unused variables (prompts service/test)
- Fix SIM105: try/except/pass → contextlib.suppress (conftest files)
- Auto-fix I001/F401/UP017 (import sort, unused imports, UTC)

mypy:
- Add TYPE_CHECKING import of User in security.py (avoids circular import)
- Add return type annotations to get_current_user/get_current_admin
- Type auth/router.py me() endpoint with User
- Fix ASGI middleware: use Message type for send callback
- Fix _json_response content type to dict[str, Any]
- Cast SQLAlchemy row attributes (analytics int(), executor append)
- Use Literal for LLMConfig provider (agents service)
- Cast float return in evals judge
- Remove unused type: ignore comments (evals service, main)
- Add None check before float comparison (evals test)

pyproject.toml:
- Add types-python-jose, types-passlib to dev deps
- Add mypy override to ignore pgvector missing stubs

All CI checks pass: ruff check ✓, mypy app ✓, pytest 231 passed (93.52%)

+++ backend/app/domains/evals/judge.py
@@ -96,7 +96,7 @@ def _cosine(a: list[float], b: list[float]) -> float:
     norm_b = sum(y * y for y in b) ** 0.5
     if norm_a == 0 or norm_b == 0:
         return 0.0
-    return dot / (norm_a * norm_b)
+    return float(dot / (norm_a * norm_b))
 
 
 __all__ = ["JudgeResult", "judge_contains", "judge_exact", "judge_llm", "judge_semantic"]
